+++ app/gui2/e2e/locate.ts
@@ -81,6 +81,9 @@ export function graphNodeIcon(node: Node) {
 export function selectedNodes(page: Page | Locator): Node {
   return page.locator('.GraphNode.selected') as Node
 }
+export function outputNode(page: Page | Locator): Node {
+  return page.locator('.GraphNode.outputNode') as Node
+}
 
 // === Data locators ===
 
@@ -96,6 +99,7 @@ function componentLocator<T extends string>(className: SanitizeClassName<T>) {
 }
 
 export const graphEditor = componentLocator('GraphEditor')
+export const codeEditor = componentLocator('CodeEditor')
 // @ts-expect-error
 export const anyVisualization = componentLocator('GraphVisualization > *')
 export const loadingVisualization = componentLocator('LoadingVisualization')
@@ -130,6 +134,10 @@ export function componentBrowserEntryByLabel(page: Locator | Page, label: string
 
 export function rightDock(page: Page) {
   return page.getByTestId('rightDock')
+}
+
+export function bottomDock(page: Page) {
+  return page.getByTestId('bottomDock')
 }
 
 export const navBreadcrumb = componentLocator('NavBreadcrumb')