+++ lib/scala/searcher/src/main/scala/org/enso/searcher/memory/InMemorySuggestionsRepo.scala
@@ -353,4 +353,8 @@ class InMemorySuggestionsRepo(implicit ec: ExecutionContext)
   }
 
   def close(): Unit = {}
+
+  override def toString(): String = {
+    "InMemorySuggestionsRepo"
+  }
 }